requirements.txt
Django>=5.1.7
django-crispy-forms>=2.0
python-dotenv>=1.0.0
Pillow>=10.0.0
crispy-tailwind==1.0.3
djangorestframework>=3.16.0
django-environ>=0.12.0
psycopg2-binary>=2.9.10
